+++ tests/unit_tests/common/test_common_actor.py
@@ -158,7 +158,7 @@ class TestActor:
         )
 
         # Act, Assert
-        assert actor.state == ComponentState.INITIALIZED
+        assert actor.state == ComponentState.POST_INITIALIZED
         assert actor.is_initialized
 
     def test_register_warning_event(self):
@@ -790,7 +790,7 @@ class TestActor:
 
         # Assert
         assert "on_reset" in actor.calls
-        assert actor.state == ComponentState.INITIALIZED
+        assert actor.state == ComponentState.POST_INITIALIZED
 
     def test_dispose(self):
         # Arrange